# openapi/api.yaml
...
openapi: 3.0.3
info:
  title: SurenSaaS API
  version: 1.0.0
  description: API multi-PME pour construction, nettoyage et diagnostic énergétique
servers:
  - url: https://back-xxx.run.app/api/v1
    description: Production
paths:
...
  /clients:
    get:
      summary: Lister les clients
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Liste des clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Client'
        401:
          description: Non authentifié
        403:
          description: Accès non autorisé

    post:
      summary: Créer un client
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientCreate'
      responses:
        201:
          description: Client créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        401:
          description: Non authentifié
  /clients/{client_id}:
    get:
      summary: Détails d'un client
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Détails du client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        404:
          description: Client non trouvé

    put:
      summary: Modifier un client
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClientUpdate'
      responses:
        200:
          description: Client modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'

    delete:
      summary: Supprimer un client
      parameters:
        - name: client_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: org_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Client supprimé
        403:
          description: Non autorisé

  # ============================================
  # INVOICES (NOUVELLE VERSION)
  # ============================================
...
  /{org}/telegram/bots:
    get:
      summary: Lister les bots Telegram
      security:
        - bearerAuth: []
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: include_inactive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Liste des bots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelegramBot'

    post:
      summary: Créer un bot Telegram
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBotRequest'
      responses:
        201:
          description: Bot créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramBot'
        403:
          description: Non autorisé (capability admin requise)
  /{org}/telegram/bots/{bot_id}:
    get:
      summary: Détails et status d'un bot
      security:
        - bearerAuth: []
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Détails du bot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramBotStatus'

    delete:
      summary: Désactiver un bot
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        204:
          description: Bot désactivé
        403:
          description: Non autorisé

  # ============================================
  # COMPANIES (ENTREPRISES FILIALES)
  # ============================================
...
  /admin/telegram/bots:
    get:
      summary: Lister les bots Telegram disponibles
      security:
        - bearerAuth: []
      description: Nécessite le rôle admin. Retourne tous les bots configurés pour l'environnement actuel.
      responses:
        200:
          description: Liste des bots disponibles
          content:
            application/json:
              schema:
                type: object
                properties:
                  bots:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelegramBotInfo'
                  environment:
                    type: string
                    enum: [test, prod]
        403:
          description: Admin requis
...
  /webhook/construction:
    post:
      summary: Webhook Telegram Construction Bot
      description: |
        Reçoit les updates du bot Telegram Construction. 
        Cette route est publique mais vérifie le secret token si configuré.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Update Telegram standard
      parameters:
        - name: X-Telegram-Bot-Api-Secret-Token
          in: header
          required: false
          schema:
            type: string
          description: Secret token pour vérification (optionnel)
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  result:
                    type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        siret:
          type: string
          pattern: '^[0-9]{14}$'
        notes:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ClientCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        siret:
          type: string
          pattern: '^[0-9]{14}$'
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        siret:
          type: string
          pattern: '^[0-9]{14}$'
        notes:
          type: string


    # ----------------------------------------
    # SCHEMAS EXISTANTS (LEGACY)
    # ----------------------------------------
...
    TelegramBot:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bot_username:
          type: string
        bot_id:
          type: integer
        company_id:
          type: string
          format: uuid
          nullable: true
        is_active:
          type: boolean
        is_configured:
          type: boolean
        description:
          type: string
        welcome_message:
          type: string
        created_at:
          type: string
          format: date-time
    CreateBotRequest:
      type: object
      required:
        - bot_token
        - webhook_base_url
      properties:
        bot_token:
          type: string
          description: Token du bot (depuis @BotFather)
        company_id:
          type: string
          format: uuid
          description: Entreprise associée (optionnel)
        webhook_base_url:
          type: string
          description: URL de base pour les webhooks
        description:
          type: string
    TelegramBotStatus:
      type: object
      properties:
        id:
          type: string
          format: uuid
        bot_username:
          type: string
        is_active:
          type: boolean
        is_configured:
          type: boolean
        created_at:
          type: string
          format: date-time
        stats_24h:
          type: object
          properties:
            interactions:
              type: integer
            errors:
              type: integer
...
    TelegramBotInfo:
      type: object
      properties:
        bot_id:
          type: string
          description: Identifiant unique du bot (e.g., 'construction', 'audit')
          example: "construction"
        bot_name:
          type: string
          description: Nom affiché du bot
          example: "Construction"
        description:
          type: string
          description: Description des fonctionnalités
          example: "Envoi et traitement des factures de construction"
        icon:
          type: string
          description: Emoji pour l'interface
          example: "🏗️"
        username:
          type: string
          description: Username Telegram du bot
          example: "suren_construction_bot"
        configured:
          type: boolean
          description: Indique si le bot est configuré
...